Moved message event into file, let's hope it works
